requirements.txt
deepmerge
docker-image-py
pyyaml
